intial moment banners split and colour setup
