+++ packages/modules/src/modules/banners/globalNYMoment/GlobalNYMomentBanner.stories.tsx
@@ -1,0 +1,116 @@
+import React from 'react';
+import { Story, Meta } from '@storybook/react';
+import { getMomentTemplateBanner } from '../momentTemplate/MomentTemplateBanner';
+import { props } from '../utils/storybook';
+import { BannerProps, SecondaryCtaType } from '@sdc/shared/types';
+import { bannerWrapper } from '../common/BannerWrapper';
+import { brand } from '@guardian/src-foundations';
+
+export default {
+    title: 'Banners/MomentTemplate',
+    args: props,
+} as Meta;
+
+const GlobalNYBanner = bannerWrapper(
+    getMomentTemplateBanner({
+        backgroundColour: '#F1F8FC',
+        headerSettings: {
+            textColour: '#0077B6',
+        },
+        primaryCtaSettings: {
+            default: {
+                backgroundColour: '#0077B6',
+                textColour: 'white',
+            },
+            hover: {
+                backgroundColour: 'white',
+                textColour: '#0077B6',
+                border: `1px solid ${'#0077B6'}`,
+            },
+        },
+        secondaryCtaSettings: {
+            default: {
+                backgroundColour: '#F1F8FC',
+                textColour: '#0077B6',
+                border: `1px solid ${'#0077B6'}`,
+            },
+            hover: {
+                backgroundColour: '#0077B6',
+                textColour: '#F1F8FC',
+            },
+        },
+        closeButtonSettings: {
+            default: {
+                backgroundColour: '#F1F8FC',
+                textColour: brand[400],
+                border: `1px solid ${brand[400]}`,
+            },
+            hover: {
+                backgroundColour: '#0077B6',
+                textColour: brand[400],
+            },
+        },
+        highlightedTextSettings: {
+            textColour: '#F1F8FC',
+            highlightColour: '#0077B6',
+        },
+        imageSettings: {
+            mainUrl:
+                'https://media.guim.co.uk/a1087c3f7e6da4f1e97947acccdd7f0d15f327d4/0_0_142_124/140.png',
+            altText: 'Guardian logo being held up by supporters of the Guardian',
+        },
+    }),
+    'global-new-year-banner',
+);
+
+const GlobalNYTemplate: Story<BannerProps> = (props: BannerProps) => <GlobalNYBanner {...props} />;
+
+export const GlobalNY = GlobalNYTemplate.bind({});
+GlobalNY.args = {
+    ...props,
+    content: {
+        heading: 'Show your support for reader-funded journalism',
+        messageText:
+            'Fearless, investigative reporting shapes a fairer world. At the Guardian, our independence allows us to chase the truth wherever it takes us. We have no shareholders. No vested interests. Just the determination and passion to bring readers quality reporting, including groundbreaking investigations. We do not shy away. And we provide all this for free, for everyone.',
+        paragraphs: [
+            'Fearless, investigative reporting shapes a fairer world. At the Guardian, our independence allows us to chase the truth wherever it takes us. We have no shareholders. No vested interests. Just the determination and passion to bring readers quality reporting, including groundbreaking investigations.',
+            'We do not shy away. And we provide all this for free, for everyone.',
+        ],
+        highlightedText:
+            'Show your support today from just %%CURRENCY_SYMBOL%%1, or sustain us long term with a little more. Thank you.',
+        cta: {
+            text: 'Support the Guardian',
+            baseUrl: 'https://support.theguardian.com/contribute',
+        },
+        secondaryCta: {
+            type: SecondaryCtaType.Custom,
+            cta: {
+                text: 'Hear from our editor',
+                baseUrl: 'https://theguardian.com',
+            },
+        },
+    },
+    mobileContent: {
+        heading: 'As 2022 begins, will you support us?',
+        messageText:
+            'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus',
+        paragraphs: [
+            'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus',
+        ],
+        highlightedText:
+            'Show your support today from just %%CURRENCY_SYMBOL%%1, or sustain us long term with a little more. Thank you.',
+        cta: {
+            text: 'Support us',
+            baseUrl: 'https://support.theguardian.com/contribute',
+        },
+        secondaryCta: {
+            type: SecondaryCtaType.Custom,
+            cta: {
+                text: 'Learn more',
+                baseUrl: 'https://theguardian.com',
+            },
+        },
+    },
+    numArticles: 50,
+    tickerSettings: undefined,
+};
